## Problem
DRF perf changes could squeeze out perhaps an 0.25s performance increase in dashboard load (go/revenue goes from 1.5s to 1.25s)

## Changes
Replace DRF for dashboard tiles down with raw dict serializers. This speeds us up, but potentially hinders maintainability. Worth discussing.

  📊 Final Results:

  | Version          | Time per Call | Total Speedup | Main Bottleneck                        |
  |------------------|---------------|---------------|----------------------------------------|
  | DRF Baseline     | 0.306s        | 1.0x          | DRF serialization overhead             |
  | Fast Serializers | 0.129s        | 2.4x          | Database queries                       |
  | Parallel + Fast  | 0.096s        | 3.2x          | Database queries + thread coordination |

Testing on a 40 insight dashboard:(100 runs)
DRF: 30.7s 
<img width="1359" height="1108" alt="Screenshot From 2025-08-22 18-32-08" src="https://github.com/user-attachments/assets/e6f70512-d365-4e0f-b765-650d5ecc8dda" />

Fast Serializers: 12.9s
<img width="1372" height="1074" alt="Screenshot From 2025-08-22 18-36-56" src="https://github.com/user-attachments/assets/c7c3b843-ff1a-4850-a73b-20dc1556b851" />

Threaded + Fast Serializers: 9.6s
<img width="1160" height="406" alt="image" src="https://github.com/user-attachments/assets/c7ea1d70-9af5-4811-9f61-ae92a372d45e" />


To copy over the table from the other PR:
● 🎉 INCREDIBLE RESULTS! Complete Performance Comparison:

  | Test                       | Time per Call | Total Time (20 calls) | Speedup vs Baseline | Key Bottlenecks                                      |
  |----------------------------|---------------|-----------------------|---------------------|------------------------------------------------------|
  | BASELINE                   | 1.154s        | 23.1s                 | 1.0x                | Pydantic model construction (13.8s), DRF serializers |
  | to_dict Only               | 0.327s        | 6.5s                  | 🚀 3.5x             | DRF serializers (3.2s), remaining field construction |
  | to_dict + fast_serializers | 0.171s        | 3.4s                  | 🚀 6.7x             | Database queries, minimal serialization overhead     |


Looking at the SnakeViz results, **we're down to scraps now!** The heavy framework overhead is eliminated and you're left with actual business logic. Here are the next 3 optimization
  opportunities, though they're much smaller wins:

  ### 1. Database Query Optimization (~20-30% potential gain)
  **What:** `django.db.models.sql.compiler.execute_sql` and query fetching
  **Why:** Multiple small queries instead of optimized joins
  **Fix:**
  - Add `select_related()` and `prefetch_related()` to dashboard tile queries
  - Combine related model fetches into single queries
  - Review for N+1 query patterns

  ### 2. Insight Calculation Caching (~15-25% potential gain)
  **What:** `query_runner.py` and `calculate_for_query_based_insight`
  **Why:** Recalculating insights that might be cacheable
  **Fix:**
  - Improve insight result caching strategies
  - Add smarter cache invalidation
  - Pre-calculate common dashboard insights

  ### 3. JSON Serialization Micro-optimizations (~5-10% potential gain)
  **What:** The remaining `fast_serializers.py` overhead
  **Why:** Could potentially use even faster serialization
  **Fix:**
  - Profile orjson vs other serializers
  - Minimize object creation during serialization
  - Cache serialized responses when appropriate

  ## 💡 Honest Assessment

  You've achieved **85% performance improvement** already. These remaining optimizations would give you maybe **another 20-40%** on top, but with **significantly more complexity**.



## How did you test this code?
Tests.